Extracted ssm logic into secrets.py
