+++ dbt_platform_helper/providers/secrets.py
@@ -1,6 +1,10 @@
 import json
 import urllib
 
+from dbt_platform_helper.constants import CONDUIT_ADDON_TYPES
+from dbt_platform_helper.exceptions import AddonNotFoundError
+from dbt_platform_helper.exceptions import InvalidAddonTypeError
+from dbt_platform_helper.exceptions import ParameterNotFoundError
 from dbt_platform_helper.exceptions import SecretNotFoundError
 
 
@@ -34,3 +38,45 @@ def get_connection_secret_arn(ssm_client, secrets_manager_client, secret_name: s
         pass
 
     raise SecretNotFoundError(secret_name)
+
+
+def get_addon_type(ssm_client, application_name: str, env: str, addon_name: str) -> str:
+    addon_type = None
+    try:
+        addon_config = json.loads(
+            ssm_client.get_parameter(
+                Name=f"/copilot/applications/{application_name}/environments/{env}/addons"
+            )["Parameter"]["Value"]
+        )
+    except ssm_client.exceptions.ParameterNotFound:
+        raise ParameterNotFoundError
+
+    if addon_name not in addon_config.keys():
+        raise AddonNotFoundError
+
+    for name, config in addon_config.items():
+        if name == addon_name:
+            addon_type = config["type"]
+
+    if not addon_type or addon_type not in CONDUIT_ADDON_TYPES:
+        raise InvalidAddonTypeError(addon_type)
+
+    if "postgres" in addon_type:
+        addon_type = "postgres"
+
+    return addon_type
+
+
+def get_parameter_name(
+    application_name: str, env: str, addon_type: str, addon_name: str, access: str
+) -> str:
+    if addon_type == "postgres":
+        return f"/copilot/{application_name}/{env}/conduits/{_normalise_secret_name(addon_name)}_{access.upper()}"
+    elif addon_type == "redis" or addon_type == "opensearch":
+        return f"/copilot/{application_name}/{env}/conduits/{_normalise_secret_name(addon_name)}_ENDPOINT"
+    else:
+        return f"/copilot/{application_name}/{env}/conduits/{_normalise_secret_name(addon_name)}"
+
+
+def _normalise_secret_name(addon_name: str) -> str:
+    return addon_name.replace("-", "_").upper()
